Update the CMake-based build system.
To the new project.

+++ test/test.c
@@ -20,7 +20,7 @@ main(int argc, char *argv[])
 {
     int   sockfd;
     pid_t pid;
-    int   ndx;
+    long  ndx;
 
     srvinitsignals();
     sockfd = srvinitsock();
